backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
passlib==1.7.4
pydantic>=1.7.4,<2.0.0
redis>=4.2.0rc1,<5.0.0
fastapi-cache2==0.2.1
fastapi-limiter==0.1.5
loguru==0.7.2
sentry-sdk==1.38.0
prometheus-fastapi-instrumentator==6.1.0
aiosmtplib==2.0.2
fastapi-utils==0.2.1
python-multipart==0.0.6
email-validator==2.1.0.post1
bcrypt==4.0.1
websockets==12.0
asyncpg==0.29.0
alembic==1.13.0
SQLAlchemy>=1.4.0,<2.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
